Fix color printing on Linux by setting color right before print
Moved set_console_attributes() next to va_start(), since the color
is reset with the *_prefix() calls.

On Windows, the restore_console_attributes(attr) call would properly
restore the color set in the show_* calls. On Linux, however, this
does not restore the color - instead, this simply resets the color to
neutral.

Because of that, the color is only enabled for the prefix, and not the
message itself. This is fixed by setting the color after the *_prefix()
calls, right before the va_start()/vprintf() printing calls.

(Yes, it's dirty... but the Linux version doesn't implement any sort of
terminal state tracking, or even a color setting history... which would
work, but might be a bit dirty in itself, since the user or another
program may have set their own terminal colors as well. If you'd like to
play around with this and implement it properly, let me know!)