Enabled assembly timing on all platforms
On Windows, assembly timing is achieved with _ftime using the include
file <sys/timeb.h>. As it turns out, this file also exists on *nix,
with the exact same function/struct... just named ftime() instead of
a strange _ftime!

The fix? Just added the "original" ftime() when the platform isn't
Windows, and added the include for other platforms as well.