Replace broken hash table with mostly-STL functions.
This replaces the old version of `hash_t` with a subclass of
`std::unordered_map`, which can actually resize unlike the old
implementation.

The hash table API is unchanged, and we still use MurmurHash as a hash
function, but it's been updated to MurmurHash3.

As a side effect of `unordered_map`, we need to build in C++11 mode.
The C++98 approach would be to use `std::map` instead, which would
likely be a bit slower.