Removed limits on labels and defines
Thanks to the new dynamic hash table (which uses std::map), we are now
able to remove the limits on labels and defines. This removes the
parameter from hash_init(), as well as updating calls to it everywhere
else.

In short - now you can use unlimited labels and unlimited #defines!
(At least within your memory space!)