Handle overflow and INT_MIN/-1 in expressions.
This fixes a crash in expression parsing when attempting to evaluate
`INT_MIN / -1` (SIGFPE), and handles integer overflow in translation
of literal values gracefully. As a side effect, it cleans up the
parsing code for numeric literals.

Overflow in evaluation is *not* handled, because doing so is too hard in
C for me to care at this point, but that's arguably a bug because doing
so invokes undefined behaviour.