Assembler now exits with an error code upon encountering a fatal error
This has been added because not exiting when an error has been encountered
has made it much more difficult to track find bugs in my projects. Instead
of building the project (I've been using a makefile, and the error output
gets hidden most of the time) and wondering why my program will not work,
the build will immediately fail, with the error right at the bottom of the
terminal.