Fix regression in assembling files (issue #42, thanks @jacobly!)
Newer error handling code caused the assembler to output errors and
exit on its first pass. This is a problem because the first pass
will likely always error, and the issues in the first pass are meant
to be resolved in the second pass.

As a result, this change has been reverted... for now. We'll look
into a more rigid test case to figure out where SPASM should exit.

Reverted change in errors.cpp, and reverted the testcase change in
empty-macro.asm.