Fix '.echo >0' crash in first pass
When run_first_pass_line / run_first_pass_line_sec changes the ptr
to NULL, run_first_pass does not check for this and attempts to
dereference the NULL ptr, causing a crash. Adding a NULL check and
breaking out of the loops fixes this!

Fixes one of the fuzzed lines in #43.