Fix malformed addinstr crash (ref: #43)
instr->instr_size was not being checked, and as a result an
abnormally long .addinstr would result in the following for loop
modifying instr->args, and possibly other parts of the memory.

This was fixed by checking to make sure that instr->instr_size
was 8 or below, and if not, failing.

As a result of a addinstr being possibly invalid, errors.h was
updated to reflect a possibly invalid addinstr.