Add a Windows application manifest
In particular this requests UTF-8 mode for versions of Windows that support
using the UTF-8 codepage. This is useful because it allows us to use bytestrings
and assume they're always UTF-8, without needing to deal with the mess that is
Windows wide characters and UCS-2 being completely unusable for non-BMP characters.